Merge patch from JeremyF:
37-hg-private-stack

HELGRIND: by default, assume that thread stacks are thread-local. This
means that they are always initalized to be exclusively owned by their
thread (rather than virgin), and no access checks are generated for
stack-relative memory references (ie, relative off ESP or EBP). This
saves about 70% (statically) of checks on memory accesses.

This is enabled by default, but can be disabled with
--private-stacks=no


git-svn-id: svn://svn.valgrind.org/valgrind@1308 a5019735-40e9-0310-863c-91ae7b9d1cf9