Merge patch from JeremyF:
40-hg-tidstate

HELGRIND: record more information when a memory location changes
state. Now also records the thread ID and the previous state at that
point (so now a "possible race" error message tells you the moment we
entered an error state from an OK state, and where we entered that OK
state and from what).


git-svn-id: svn://svn.valgrind.org/valgrind@1311 a5019735-40e9-0310-863c-91ae7b9d1cf9