Merge revs 1.56 and 1.57:
(1.56):
speed optimisation: assuming that a tempreg not related to %esp before
is PUTed into %ESP, we do *know* that this tempreg points to %ESP afterwards,
so track it.

reduces the amount of expensive "DO_GENERIC" calls by 75% in Khtml. I've
not observed a problem with it, but consider it experimental.

(1.57):
Improved the ESP-update pass -- it's method for finding the ESP-deltas was very
fragile, and it was lucky that it worked at all.  Replaced it with a much more
robust method.  Thanks to Dirk Mueller for pointing it out.


git-svn-id: svn://svn.valgrind.org/valgrind@1889 a5019735-40e9-0310-863c-91ae7b9d1cf9