(Mostly) merged from HEAD:
Updated the README file.  It had lots of out-of-date and incorrect information
in it, much of it from 1.0.X days.  Did it in such a way that if it doesn't get
touched (and it undoubtedly won't) it won't really go out of date, eg. by
removing temporary details like version numbers, dates, details of specific
software incompatibilities.  It's much better to be vague but correct, than
precise but incorrect;  having incorrect info in a file as important as the
README is bad.  Also removed the README_KDE3_FOLKS file because it's pretty
redundant now.

Did similar, but smaller changes to README_DEVELOPERS and README_PACKAGERS.

Also updated the valgrind.spec.in file to use the new, post-1.0.X description
in the README.

Also fixed a minor omission in Addrcheck's docs.


git-svn-id: svn://svn.valgrind.org/valgrind@2090 a5019735-40e9-0310-863c-91ae7b9d1cf9