When signal routing is in use (because we are running on an older kernel that doesn't route signals to the correct threads properly) the siginfo data was not being propagated to any signal handlers installed by the client program.
This is because the main thread routes the signal to the proxy LWP by
using the tkill system call but that then appears in the proxy as a user
initiated signal and the original siginfo data is lost.

This patch adds a small queue of siginfo data for pending sigals to
each thread's state so that when the proxy LWP passes the signal back
to the main thread the relevant siginfo data can be recovered and passed
on to any signal handlers. Thix fixes bug 91325.


git-svn-id: svn://svn.valgrind.org/valgrind@2773 a5019735-40e9-0310-863c-91ae7b9d1cf9