Simplify the implementation of m_trampoline and the stuff that refers to it.  Now there are no more offsets and no copying of code into the stack.  We just redirect directly to entry points in m_syscalls.S.
This will mess up pointercheck, since the redirect targets are now in
Valgrind's address space, not the client's.  But pointercheck is hosed
anyway, and I'd rather back off to something simple whilst ppc32 is
stabilised.  When the address space management stuff is overhauled
then pointercheck may or may not get reinstated, and if it does then
the trampoline stuff will need revisiting.



git-svn-id: svn://svn.valgrind.org/valgrind@3977 a5019735-40e9-0310-863c-91ae7b9d1cf9