First cut at compressed V bits.  Almost all the fast cases are done, except the STOREV8 fast case is commented out because there's a bug in it (demonstrated with memcheck/tests/fprw) that I haven't yet found.
It passes all the regtests on my x86/Linux box.  I haven't tested it on an
AMD64 or PPC32 so some of the 64-bit or big-endian stuff might be broken.

It's currently as fast or faster than the trunk on the few programs I've
measured.  We're still a lot slower than 2.4.0, though.


git-svn-id: svn://svn.valgrind.org/valgrind@5218 a5019735-40e9-0310-863c-91ae7b9d1cf9