Change core-tool interface to take account of variable location reading machinery:
VG_(needs_data_syms) is gone.  Data symbols are always read.

It is replaced by VG_(needs_var_info).  This asks for variable info to
be read.  The only tools to use this are Helgrind and DRD.

The user can force loading of variable info, even if the tool
doesn't request it.  This is useful for eg making Memcheck produce
better error messages.  A new flag is provided for this:

--read-var-info=yes|no    read variable type & location info? [no]



git-svn-id: svn://svn.valgrind.org/valgrind@7486 a5019735-40e9-0310-863c-91ae7b9d1cf9