Change the meaning of origin tags to be as described in Nethercote et al's paper "Tracking Bad Apples: ...", sec 5.2 "Representing Origins".
Origin tags are still 32-bit values.  However, they are now abstract
handles to ExeContexts made at the time of the allocation.  For heap
blocks, the stack trace at allocation is used.  For stack allocations,
a trivial 1-element stack trace is used, for performance reasons.

Most of the changes are to do with the relatively complex mechanisms
to do with handling stack allocation/deallocation quickly.

These changes are prerequisites needed to make this origin tracking
scheme work on 64-bit targets.




git-svn-id: svn://svn.valgrind.org/valgrind@7838 a5019735-40e9-0310-863c-91ae7b9d1cf9