Add a secondary, inclusive, write-back, dynamically allocated, free range, organically grown, secondary cache for storing origin tags and superfluous adjectives.  When a line containing useful information is ejected from the main origin tag cache, it is placed in the secondary cache.  Later, if we miss on the primary cache, the secondary is searched for the line (obviously).
This is believed to fix the phenomenon in which the origin tracking
machinery fails to show an origin as a result of capacity limitations
in the main origin cache.  Assuming that this is all working correctly
(cough, cough), the process will eat up all available memory to store
origin tags, if necessary, in preference to losing any of them.



git-svn-id: svn://svn.valgrind.org/valgrind@7913 a5019735-40e9-0310-863c-91ae7b9d1cf9