This commit adds an experimental tool, Ptrcheck, previously known as Annelid.  Annelid is a tool for bounds checking, written by Nick Nethercote, as contained in
http://www.valgrind.org/downloads/njn/valgrind-annelid.tar.bz2.

This commit contains those sources ported to the current (3.x)
Valgrind framework.

Note this is a work in progress and the current port runs only
the simplest programs (if at all), although the original Annelid
tool worked quite reliably.

The ideas behind the tool are described in detail in

Bounds-Checking Entire Programs Without Recompiling.
  Nicholas Nethercote and Jeremy Fitzhardinge.  Informal
  Proceedings of the Second Workshop on Semantics, Program
  Analysis, and Computing Environments for Memory
  Management (SPACE 2004), Venice, Italy, January 2004.

  http://www.valgrind.org/docs/bounds-checking2004.ps

and also chapter 4 of Nick's PhD thesis, 

Dynamic Binary Analysis and Instrumentation.

  Nicholas Nethercote.
  PhD Dissertation, University of Cambridge, November 2004.

  http://www.valgrind.org/docs/phd2004.pdf



git-svn-id: svn://svn.valgrind.org/valgrind@8352 a5019735-40e9-0310-863c-91ae7b9d1cf9