Introduce a new type, PtrdiffT.  Replace lots of uses of OffT (all those that are memory offsets) with PtrdiffT;  OffT should only be used for file sizes and offsets.
Change Off64T from a ULong to a Long, as it should be.  Replace some uses
of ULong in the address space manager with Off64T to match.

Also add a comment explaining the meanings of the basic types like Addr,
OffT, SizeT, etc.

Also fix the prototype for VG_(pread) -- the last arg is an OffT, not an
Int.



git-svn-id: svn://svn.valgrind.org/valgrind@8959 a5019735-40e9-0310-863c-91ae7b9d1cf9