Fill in slots in higher-level logic needed to handle delivery of signals to threads waiting in syscalls on Darwin.  This really just means implementing VG_UCONTEXT_SYSCALL_SYSRES.
Add many comments, explaining the meaning of
VG_UCONTEXT_SYSCALL_SYSRES, and explaining the logic in
async_sighandler.

Got rid of VG_UCONTEXT_SYSCALL_NUM.  It's not needed for anything
except debug printing, and in any case the notion that the syscall
number can reliably (portably) be extracted from the interrupting
signal's ucontext is simply incorrect.




git-svn-id: svn://svn.valgrind.org/valgrind@9465 a5019735-40e9-0310-863c-91ae7b9d1cf9