Separate VG_ and ML_ definitions for asm files vs. C files.  This lets us paper over the fact that C names in Darwin get an underscore prepended to them, and get rid of the _VG_ and _ML_ macros that were being used on Darwin.
As part of this, pub_{core,tool}_basics_asm.h only now are included in asm
files, not C files.  This is no great loss as they only included the VG_ and
ML_ definitions.

This allowed m_cpuid-{x86,amd64}-darwin.S to be removed;  m_cpuid.S can now
be used as is on Darwin.

Also removed an unnecessary bit of conditional compilation in fpu-28-108.S
and fixed VG_SYM_ASM in tests/asm.h.


git-svn-id: svn://svn.valgrind.org/valgrind@9600 a5019735-40e9-0310-863c-91ae7b9d1cf9