Change the way VG_(message) behaves w.r.t. newlines, to make it more like printf.
Prior to this commit, it would print the ==PID== preamble before the
message, and a newline after it, and if the message itself contained a
newline then the result was bad -- those new lines would not start
with a preamble.

This commit adds two changes in behaviour:

* if the message contains a newline, then the preamble will be printed
  after each one.  Hence it is now possible to use VG_(message) to
  print multi-line messages correctly.

* VG_(message) no longer implicitly adds a newline at the end of the
  message -- callers must do that themselves, explicitly, exactly as
  with printf.  Benefit is that is is now possible to use multiple
  calls to VG_(message) to build up a single line of output.

VG_(message) now buffers output more aggressively than it did before,
to save on write syscalls, although it will immediately flush any any
time it encounters a newline.  As a result of this buffering it is
necessary to add a VG_(message_flush) routine, to flush any buffered
output.




git-svn-id: svn://svn.valgrind.org/valgrind@9808 a5019735-40e9-0310-863c-91ae7b9d1cf9