When looking for the text section in a PE executable ignore any code section which is marked as uninitialised.
This can happen if you have incremental linking enabled in Visual
Studio, which causes a .textbss section to be added before the real
text section. We were picking up that .textbss section and using it to
compute the avma and bias for the code which was giving completely the
wrong results.


git-svn-id: svn://svn.valgrind.org/valgrind@10394 a5019735-40e9-0310-863c-91ae7b9d1cf9